# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: decant daemon API
  version: "1"
  description: >
    Read API for the decant daemon (Plan 3). All endpoints live under `/api/v1`,
    return the `{data, meta, errors}` envelope, and perform server-side
    aggregation only — raw rows are never shipped except the messages/blocks of a
    single session detail. Lists use opaque cursor pagination
    (`base64({rowid, sort_key})`). Every endpoint except `/health` requires a
    bearer token and is bound to loopback with a Host-header allowlist
    (see the architecture design, §4–§5).

    This document is the contract of record. The recommendations endpoints
    (`/recommendations*`) are defined in a later plan and are intentionally
    omitted here.
servers:
  - url: http://127.0.0.1:4577/api/v1
    description: Local daemon (default port)
security:
  - bearerAuth: []
...
paths:
...
  /analytics/summary:
    get:
      tags: [analytics]
      summary: Totals scoped to filters
      parameters:
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
        - $ref: "#/components/parameters/Tool"
        - $ref: "#/components/parameters/Model"
        - $ref: "#/components/parameters/Project"
        - $ref: "#/components/parameters/Outcome"
        - $ref: "#/components/parameters/WorkType"
      responses:
        "200":
          description: Aggregate totals.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/Totals" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /analytics/activity:
    get:
      tags: [analytics]
      summary: Activity histograms (local time)
      description: >
        Session counts bucketed by local hour (24) and weekday (7, Sunday=0),
        computed with SQLite `strftime(..., 'localtime')`. Includes the server's
        UTC offset label and the peak bucket indices.
      parameters:
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
        - $ref: "#/components/parameters/Tool"
        - $ref: "#/components/parameters/Model"
        - $ref: "#/components/parameters/Project"
      responses:
        "200":
          description: Activity histograms.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/Activity" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /analytics/model-sparklines:
    get:
      tags: [analytics]
      summary: Per-model daily counts on a shared day axis
      parameters:
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
        - $ref: "#/components/parameters/Tool"
        - $ref: "#/components/parameters/Model"
        - $ref: "#/components/parameters/Project"
      responses:
        "200":
          description: Sparkline series.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/ModelSparklines" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /analytics/files:
    get:
      tags: [analytics]
      summary: File hotspots (or per-extension language lens)
      description: >
        Which files agents touch most, with per-operation counts
        (read/edit/write/delete), distinct session coverage, and last touch.
        Backed by `file_ref` rows extracted deterministically at ingest (Claude
        Read/Edit/Write/NotebookEdit paths; Codex apply_patch headers — Codex
        shell reads are not mined, so its read counts are conservative).
        Top-N by total operations; no cursor.
      parameters:
        - name: group
          in: query
          description: Aggregation key — project-relative path or file extension.
          schema:
            type: string
            enum: [path, ext]
            default: path
        - name: op
          in: query
          description: Keep only one operation kind.
          schema:
            type: string
            enum: [read, edit, write, delete]
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
        - $ref: "#/components/parameters/Tool"
        - $ref: "#/components/parameters/Model"
        - $ref: "#/components/parameters/Project"
        - $ref: "#/components/parameters/Outcome"
        - $ref: "#/components/parameters/WorkType"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Hotspot rows, ordered by total operations desc.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/FileStatRow" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /analytics/now:
    get:
      tags: [analytics]
      summary: Glanceable "now" snapshot
      description: >
        One cheap call for glanceable clients (menu bar, dashboard header):
        today's totals (local-time day bounds), the sessions whose source files
        are being written right now (live-session tracker; active = a write in
        the last 120s), and sync state. Live entries are joined to their archive
        row when already ingested; a brand-new session may briefly carry only
        `tool` + `source_path`.
      responses:
        "200":
          description: The snapshot.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/Now" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        32-byte hex token from `~/.decant/daemon.token` (or `DECANT_DAEMON_TOKEN`).
        Required on every `/api/*` request except `/health`.
...
  parameters:
    From:
      name: from
      in: query
      description: Inclusive lower bound on session start date (YYYY-MM-DD).
      schema: { type: string, format: date }
    To:
      name: to
      in: query
      description: Inclusive upper bound on session start date (YYYY-MM-DD). The whole day is included.
      schema: { type: string, format: date }
    Tool:
      name: tool
      in: query
      description: Exact tool match (e.g. `claude_code`, `codex`).
      schema: { type: string }
    Model:
      name: model
      in: query
      description: Exact model match.
      schema: { type: string }
    Project:
      name: project
      in: query
      description: >
        Project path. A resolved root path matches the root checkout plus every
        worktree rolled up under it; any other value falls back to an exact
        project-path match.
      schema: { type: string }
...
    Outcome:
      name: outcome
      in: query
      description: >
        Heuristic session outcome (computed at ingest). Unknown values are a
        `400`, never a silent empty match.
      schema:
        type: string
        enum: [completed, failed, abandoned]
    WorkType:
      name: work_type
      in: query
      description: >
        Heuristic session work type (computed at ingest from the first prompt's
        intent keywords, falling back to tool mix).
      schema:
        type: string
        enum: [debugging, feature, refactor, research, ops]
    Limit:
      name: limit
      in: query
      description: Page size. Defaults to 50, clamped to 200.
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
  schemas:
...
    Pagination:
      type: object
      properties:
        next_cursor:
          type: [string, "null"]
          description: Cursor for the next page, or null if this is the last page.
        has_more: { type: boolean }
        total_count:
          type: integer
          description: >
            Total rows matching the filters (ignores the cursor); stable across
            pages. Omitted for endpoints where a total would require a second
            full scan (search).
        page_size: { type: integer }
      required: [has_more, page_size]
    Meta:
      type: object
      properties:
        pagination: { $ref: "#/components/schemas/Pagination" }
        filters_applied:
          type: object
          additionalProperties: true
          description: The subset of filters that were applied (set keys only).
        timestamp: { type: string, format: date-time }
      required: [timestamp]
    Envelope:
      type: object
      description: The standard success envelope. `data` shape varies by endpoint.
      properties:
        data: {}
        meta: { $ref: "#/components/schemas/Meta" }
        errors:
          type: array
          items: {}
          description: Always empty on success; reserved for partial-result warnings.
      required: [data, meta, errors]
...
    ApiError:
      type: object
      properties:
        code:
          type: string
          description: Stable machine-readable error code.
          enum:
            - INVALID_REQUEST
            - INVALID_FILTER
            - INVALID_QUERY
            - NOT_FOUND
            - INTERNAL
        message: { type: string }
        hint: { type: [string, "null"] }
        details: { type: [object, "null"], additionalProperties: true }
        request_id: { type: string, description: Correlation id for logs.}
      required: [code, message, request_id]
    ErrorBody:
      type: object
      description: >
        Error response. Mirrors the envelope shape (data null, errors []) and adds
        a top-level `error` object.
      properties:
        data: { type: "null" }
        meta: { $ref: "#/components/schemas/Meta" }
        errors:
          type: array
          items: {}
        error: { $ref: "#/components/schemas/ApiError" }
      required: [error]
...
    FileStatRow:
      type: object
      description: One file-hotspot (or extension) rollup row.
      properties:
        key:
          type: string
          description: Project-relative path (absolute fallback) for `group=path`; extension for `group=ext`.
        project:
          type: [string, "null"]
          description: Owning project path (`group=path` only).
        reads: { type: integer }
        edits: { type: integer }
        writes: { type: integer }
        deletes: { type: integer }
        sessions: { type: integer, description: Distinct sessions touching this key. }
        last_touched_at: { type: [string, "null"], format: date-time }
      required: [key, reads, edits, writes, deletes, sessions]
    NowSession:
      type: object
      description: A currently-active session (its source file changed within the activity window).
      properties:
        tool: { type: string }
        source_path: { type: string }
        idle_seconds: { type: integer, description: Seconds since the file last changed. }
        title: { type: [string, "null"] }
        project: { type: [string, "null"] }
        model: { type: [string, "null"] }
        started_at: { type: [string, "null"], format: date-time }
      required: [tool, source_path, idle_seconds]
    Now:
      type: object
      description: Glanceable snapshot for `/analytics/now`.
      properties:
        today: { $ref: "#/components/schemas/Totals" }
        active_sessions:
          type: array
          items: { $ref: "#/components/schemas/NowSession" }
        last_sync_at: { type: [string, "null"], format: date-time }
        sync_in_progress: { type: boolean }
      required: [today, active_sessions, sync_in_progress]
...
    Totals:
      type: object
      properties:
        sessions: { type: integer }
        messages: { type: integer }
        tool_calls: { type: integer }
        input_tokens: { type: integer }
        output_tokens: { type: integer }
        cache_read_tokens: { type: integer }
        cache_creation_tokens: { type: integer }
        estimated_cost_usd: { type: number }
      required:
        - sessions
        - messages
        - tool_calls
        - input_tokens
        - output_tokens
        - cache_read_tokens
        - cache_creation_tokens
        - estimated_cost_usd
...
    Activity:
      type: object
      properties:
        by_hour:
          type: array
          items: { type: integer }
          minItems: 24
          maxItems: 24
          description: Session counts per local hour, index 0..23.
        by_weekday:
          type: array
          items: { type: integer }
          minItems: 7
          maxItems: 7
          description: Session counts per local weekday, index 0..6 (0 = Sunday).
        timezone:
          type: string
          description: Server local UTC-offset label, e.g. "UTC-07:00".
        peak_hour: { type: [integer, "null"], description: Index of the busiest hour, or null if no data. }
        peak_weekday: { type: [integer, "null"], description: Index of the busiest weekday, or null if no data. }
      required: [by_hour, by_weekday, timezone]
    ModelSparklines:
      type: object
      properties:
        models:
          type: object
          additionalProperties:
            type: array
            items: { type: integer }
          description: Map of model -> per-day session counts, aligned to `days`.
        days:
          type: array
          items: { type: string, format: date }
          description: Sorted distinct days forming the shared x-axis.
      required: [models, days]
...
  responses:
    BadRequest:
      description: Invalid filter, parameter, or cursor.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorBody" }
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorBody" }
...